requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
dj-database-url==1.0.0
Django==4.1.2
django-cors==0.1
django-cors-headers==3.13.0
django-on-heroku==1.1.2
djangorestframework==3.14.0
drf-spectacular==0.24.2
drf-yasg==1.21.4
future==0.18.2
gunicorn==20.1.0
httmock==1.4.0
httpie==3.2.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.0
MarkupSafe==2.1.1
multidict==6.0.2
openapi-codec==1.3.2
packaging==21.3
Pillow==9.2.0
postgres==4.0
psycopg2-binary==2.9.4
psycopg2-pool==1.1
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.2
PySocks==1.7.1
pytz==2022.4
PyYAML==6.0
repo==0.3.0
requests==2.28.1
requests-toolbelt==0.10.1
rich==12.6.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
serializers==0.2.4
simplejson==3.17.6
sqlparse==0.4.3
tzdata==2022.5
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
